Fix enum usage

+++ Source/Client/Forms/Editor_Map.cs
@@ -1799,7 +1799,7 @@ namespace Client
 
         public static void MapEditorFillLayer(MapLayer layer, byte theAutotile = 0, byte tileX = 0, byte tileY = 0)
         {
-            GameLogic.Dialogue("Map Editor", "Fill Layer: " + layer.ToString(), "Are you sure you wish to fill this layer?", (byte)DialogueType.FillLayer, (byte)DialogueStyle.YesNo, GameState.CurLayer, GameState.CurAutotileType, tileX, tileY, (int)(Instance?.sldTileSet.Value ?? 1));
+            GameLogic.Dialogue("Map Editor", "Fill Layer: " + layer.ToString(), "Are you sure you wish to fill this layer?", DialogueType.FillLayer, DialogueStyle.YesNo, GameState.CurLayer, GameState.CurAutotileType, tileX, tileY, (int)(Instance?.sldTileSet.Value ?? 1));
         }
 
         public static void MapEditorEyeDropper()
